# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth": {
            "post": {
                "description": "auth user by login and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Authentification",
                "parameters": [
                    {
                        "description": "Account data",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.WebUserAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Msg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "List of all users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protected"
                ],
                "summary": "User list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Msg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    }
                }
            },
            "post": {
                "description": "creating user using data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create New User",
                "parameters": [
                    {
                        "description": "User create data",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.WebUserCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Msg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    }
                }
            }
        },
        "/users/:id": {
            "get": {
                "description": "User data by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protected"
                ],
                "summary": "User by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UserLoadResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete user by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protected"
                ],
                "summary": "Delete user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Msg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update user data by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protected"
                ],
                "summary": "Update user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateUserResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.Err"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.WebUserAuth": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "login"
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "password"
                }
            }
        },
        "dto.WebUserCreate": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "login",
                "password",
                "second_name"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@email.com"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "first_name"
                },
                "login": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "login"
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "password"
                },
                "second_name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "second_name"
                }
            }
        },
        "dto.WebUserLoad": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "login",
                "second_name"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@email.com"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "first_name"
                },
                "id": {
                    "type": "integer",
                    "minimum": 0
                },
                "login": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "login"
                },
                "second_name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "second_name"
                }
            }
        },
        "handler.Err": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "message text"
                },
                "status": {
                    "type": "string",
                    "example": "status text"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handler.Msg": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "message text"
                },
                "status": {
                    "type": "string",
                    "example": "status text"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "handler.UpdateUserResp": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "user": {
                    "$ref": "#/definitions/dto.WebUserLoad"
                }
            }
        },
        "handler.UserLoadResp": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "user": {
                    "$ref": "#/definitions/dto.WebUserLoad"
                }
            }
        }
    }
}